Start cleaning up test code
