+++ pyuvsim/tests/__init__.py
@@ -1,0 +1,24 @@
+# -*- mode: python; coding: utf-8 -*
+# Copyright (c) 2018 Radio Astronomy Software Group
+# Licensed under the 3-clause BSD License
+
+from __future__ import absolute_import, division, print_function
+
+import numpy as np
+
+
+# functions used by many tests
+def compare_dictionaries(dic1, dic2):
+    """
+        Recursively compare two dictionaries.
+    """
+    compare = True
+    for k in dic1.keys():
+        if isinstance(dic1[k], dict):
+            compare *= compare_dictionaries(dic1[k], dic2[k])
+        else:
+            if isinstance(dic1[k], float):
+                compare *= np.isclose(dic1[k], dic2[k], atol=1e-5)
+            else:
+                compare *= (dic1[k] == dic2[k])
+    return bool(compare)
